+++ src/main/java/com/stratio/mesos/auth/RedirectionInterceptor.java
@@ -9,8 +9,6 @@
 package com.stratio.mesos.auth;
 
 import okhttp3.Interceptor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -20,8 +18,6 @@ import java.util.List;
  * Created by shernandez on 10/03/17.
  */
 public class RedirectionInterceptor implements Interceptor {
-    Logger LOG = LoggerFactory.getLogger(RedirectionInterceptor.class);
-
     List<String> locationHistory = new ArrayList<>();
 
     @Override