+++ src/main/java/com/stratio/mesos/http/CookieInterceptor.java
@@ -10,8 +10,6 @@ package com.stratio.mesos.http;
 
 import okhttp3.Interceptor;
 import okhttp3.Response;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -21,8 +19,6 @@ import java.util.List;
  * Created by shernandez on 9/03/17.
  */
 public class CookieInterceptor implements Interceptor {
-    Logger LOG = LoggerFactory.getLogger(CookieInterceptor.class);
-
     private List<String> cookieHistory = new ArrayList<>();
 
     @Override